Move top-level services to app package
