+++ frontend/src/app/authentication/gate.guard.ts
@@ -1,7 +1,7 @@
 import { inject } from "@angular/core";
 import { ActivatedRouteSnapshot, CanActivateFn, Router, RouterStateSnapshot } from "@angular/router";
 import { map } from "rxjs";
-import { AuthenticationService } from "./authentication.service";
+import { AuthenticationService } from "../authentication.service";
 
 export const isAuthenticated: CanActivateFn = (route, state) => {
     const auth = inject(AuthenticationService);
